+++ index.js
@@ -1,22 +1,25 @@
 require("dotenv").config()
+const logger = require("my-custom-logger")
 
 const server = require("./app/Server")
 
 server.start()
-    .then(() => console.log("Aggregation API successfully started"))
+    .then(() => logger.info("Aggregation API successfully started"))
     .catch((e) => {
-        console.error("Failed to start Aggregation API. " + e)
+        logger.error("Failed to start Aggregation API. " + e)
         process.exit(1)
-    });
-
-const exitHandler = (event) => {
-    server.stop()
-        .then(() => {
-            console.log("Aggregation API has been stopped")
-        }).catch((e) => {
-        console.error("Failed to stop Aggregation API." + e)
-        console.error(e)
     })
-};
+
+const exitHandler = () => {
+    server
+        .stop()
+        .then(() => {
+            logger.info("Aggregation API has been stopped")
+        })
+        .catch((e) => {
+            logger.error("Failed to stop Aggregation API." + e)
+            logger.error(e)
+        })
+}
 
 process.on("exit", exitHandler)