+++ CHANGELOG.md
@@ -5,3 +5,9 @@
 # 1.0.1 - Hotfix
 
 * Listen Koa2 on all interfaces
+
+# 1.0.2 - Release
+
+* Register event method
+* Refactor
+* Centralized logging